requirements.txt
fastapi
uvicorn[standard]
ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.2.0
python-multipart>=0.0.9
numpy
opencv-python-headless
pytesseract